Minor refactoring tasks
