+++ v-next/hardhat/src/internal/cli/help/get-global-help-string.ts
