+++ v-next/hardhat/src/internal/cli/help/get-help-string.ts
