--- v-next/hardhat/src/internal/cli/init/constants.ts
@@ -1,1 +1,0 @@
-export const HARDHAT_NAME = "Hardhat";
